LayoutTests/platform/mac/accessibility/stale-table-rows-expected.txt

A B C
This tests that when a table has its DOM changed, all the table method still return the correct data.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rows is grid.rowCount
PASS columns is grid.columnCount
PASS successfullyParsed is true

TEST COMPLETE
